Reorganize functions into seaice package
